added docstrings and such
